be-feat: add exception filter to translate domain error to http response

+++ apps/backend/src/modules/schedule/application/usecases/delete/delete-schedule.usecase.ts
@@ -1,9 +1,16 @@
 import { ScheduleRepository } from 'src/modules/schedule/domain/schedule.repository';
+import { ScheduleNotFoundError } from '../../../domain/errors/schedule-not-found.error';
 
 export class DeleteScheduleUseCase {
   constructor(private scheduleRepository: ScheduleRepository) {}
 
   async execute(id: string) {
+    const currentSchedule = await this.scheduleRepository.getById(id);
+
+    if (!currentSchedule) {
+      throw new ScheduleNotFoundError(id);
+    }
+
     return await this.scheduleRepository.delete(id);
   }
 }
